through test05

+++ phase3/libuser.h
@@ -13,7 +13,7 @@ extern void Terminate(long status);
 extern void GetTimeofDay(int *tod);
 extern void CPUTime(int *cpu);
 extern void GetPID(int *pid);
-extern int  SemCreate(int value, int *semaphore);
+extern int  SemCreate(long value, int *semaphore);
 extern int  SemP(int semaphore);
 extern int  SemV(int semaphore);
 extern int  SemFree(int semaphore);
